🔥 unify news in one file
