+++ news.py
@@ -3,7 +3,12 @@ from unicodedata import normalize
 
 import requests
 from bs4 import BeautifulSoup as bs
+from dateutil import parser as ps
 
+from const import NEWS_URL_BM
+
+
+# main section: goes to the given url and scrapes
 
 def scrape_articles(url: str):
     parser = 'lxml'
@@ -18,6 +23,43 @@ def scrape_articles(url: str):
     content = response.text.encode('utf-8')
     soup = bs(content, parser)
     return soup
+
+
+# sources section: the article sources will be in this section organized by function names
+
+def bizmandu():
+    source = 'bizmandu'
+    first_word = "काठमाडौं।"
+    soup = scrape_articles(NEWS_URL_BM)
+    container = soup.find("ul", attrs={'class': "uk-list"}).find_all("li")
+    articles = []
+
+    for article in container:
+        # dont touch! just witness the magic
+        url = article.find('a')['href']
+        title = article.find('h3').text.strip()
+        paragraph = article.find('p', text=re.compile(
+            first_word), _class=False, id=False)
+        desc_one = paragraph.get_text().split(first_word)[1].strip()
+        desc_two = paragraph.findNext('p').text.strip()
+        description = f'{desc_one} {desc_two}'
+        raw_date = article.find(
+            'p', {'class': 'uk-article-meta'}).text.strip()
+        date = ps.parse(raw_date.split('बिजमाण्डू')[1].strip())
+
+        the_article = {
+            "date_published": date,
+            "description": fix_last_dharko(bleach(description)),
+            "image_url": None,
+            "lang": "nepali",
+            "source": source,
+            "title": bleach(title),
+            "url": url,
+        }
+
+        articles.append(the_article)
+
+    return articles
 
 
 def bleach(given_text: str) -> str:
